+++ sepes/__init__.py
@@ -15,7 +15,7 @@
 from sepes._src.backend import backend_context
 from sepes._src.code_build import autoinit, field, fields
 from sepes._src.tree_base import TreeClass
-from sepes._src.tree_index import AtIndexer, at
+from sepes._src.tree_index import at
 from sepes._src.tree_mask import is_masked, is_nondiff, tree_mask, tree_unmask
 from sepes._src.tree_pprint import tree_diagram, tree_repr, tree_str, tree_summary
 from sepes._src.tree_util import bcmap, leafwise, value_and_tree
@@ -37,7 +37,6 @@ __all__ = (
     "tree_unmask",
     "tree_mask",
     # indexing utils
-    "AtIndexer",
     "at",
     # tree utils
     "bcmap",